+++ src/widgets/averageNumber/averageNumber.ts
@@ -26,9 +26,10 @@ export class AverageNumber extends Chart {
 
             const currColor = this.getColor(data.dataSets[0].settings, 'color-yellow');
             const prevColor = this.getColor(data.dataSets[1].settings, 'color-grey');
+            const backgroundStyle = this.getBackground(this.getWidgetSetting('backgroundColor'));
 
             const str = `
-                <div class='${s["widget"]}'>
+                <div class='${s["widget"]}' style="${backgroundStyle}">
                     <div class='${s["row"]}'>
                         <div class='${s["col"]} ${s["col-100"]}'>
                             <div class="${w['title']}">