+++ src/widgets/static/settings.ts
@@ -4,6 +4,7 @@ import settingsPresets from "../../widgetSettings/settingsPresets";
 export const settings: IWidgetSettings = makeSettings({
     settings: [
         ...settingsPresets.common,
+        ...settingsPresets.background,
     ],
     dataSet: {
         initDataSets: [{viewType: 'STATIC'}],