+++ src/widgets/table/settings.ts
@@ -4,7 +4,7 @@ import settingsPresets from "../../widgetSettings/settingsPresets";
 
 export const settings: IWidgetSettings = makeSettings({
     settings: [
-        ...settingsPresets.common,
+        ...settingsPresets.common
     ],
     dataSet: {
         initDataSets: [{viewType: 'TABLE'}],